% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraerr.R
\name{paraerr}
\alias{paraerr}
\title{A function to test parameter in function surlogrank}
\usage{
paraerr(x, y, z)
}
\arguments{
\item{x}{time variable}

\item{y}{status variable}

\item{z}{group variable}
}
\value{
1 or 0
}
\description{
Whether the parameter lengths are equal,time variable input is of numeric type,status variable with only 0,1,group only two elements
}
