% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{permut}
\alias{permut}
\title{function that perform the log-rank test using the large-sample approximation}
\usage{
permut(time, status, group)
}
\arguments{
\item{time}{time variable,a numeric vector}

\item{status}{status variable,only conclude 0 and 1 ,suggest live or dead}

\item{group}{group variable,only divided to 2 groups of patients with acute myelogenous leukaemia}
}
\value{
an object that contains the test statistic, the p-value, and the method
}
\description{
use four statistics:Mantel-Haenszel Logrank Test;Cox-Mantel Logrank Test;Generalized Wilcoxon;Tarone and Ware
}
