% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_MetropolisL}
\alias{rw_MetropolisL}
\title{Random walk Metropolis using Rcpp}
\usage{
rw_MetropolisL(sigma, x0, m)
}
\arguments{
\item{sigma}{the variance}

\item{x0}{the inial location}

\item{m}{the total steps}
}
\value{
a list conclude x and k \code{n}
}
\description{
Implement the random walk version of the Metropolis using Rcpp
}
