% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surlogrank.R
\name{surlogrank}
\alias{surlogrank}
\title{function that perform the log-rank test using the large-sample approximation and the permutation test}
\usage{
surlogrank(time, status, group, permutation)
}
\arguments{
\item{time}{time variable,a numeric vector}

\item{status}{status variable,only conclude 0 and 1 ,1：fail or 0：at risk}

\item{group}{group variable,only divided to 2 groups of patients with acute myelogenous leukaemia}

\item{permutation}{TRUE or FLASE ,use permutation or not}
}
\value{
an object that contains the test statistic, the p-value, and the method
}
\description{
use four statistics:Mantel-Haenszel Logrank Test;Cox-Mantel Logrank Test;Generalized Wilcoxon;Tarone and Ware
}
